feat(core): add DECLRMM mode 69 plumbing (§08.3)
Widen TermMode from u32 to u64 to accommodate bit 32, then add the
full vertical slice for DEC Left Right Margin Mode (DECLRMM):

- VTE: NamedPrivateMode::LeftRightMargin = 69 variant + PrivateMode::new mapping
- TermMode: LEFT_RIGHT_MARGIN = 1 << 32 flag
- helpers: named_private_mode_flag exhaustive match arm
- modes: DECSET/DECRST handler arms (insert/remove flag)
- Mode reporting: DECRQM works automatically via the flag mapping

Cascade: AtomicU32→AtomicU64 for mode_cache across oriterm_mux pane,
io_thread, domain; wire protocol snapshot.modes u32→u64; MuxBackend
trait pane_mode returns u64; all test call sites updated.

Tests: mode_69_set, mode_69_reset, mode_69_decrqm_reports, exhaustive
decset_decrst_flag_sync updated, all_flags_are_distinct updated,
term_mode_size guard updated to 8 bytes.

+++ oriterm_mux/src/protocol/snapshot.rs
@@ -203,7 +203,7 @@ pub struct PaneSnapshot {
     /// | 24  | `REVERSE_WRAP` (mode 45) |
     /// | 25  | `MOUSE_URXVT` (mode 1015) |
     /// | 26  | `MOUSE_X10` (mode 9) |
-    pub modes: u32,
+    pub modes: u64,
     /// Number of scrollback rows above the viewport.
     pub scrollback_len: u32,
     /// Current scroll position (0 = bottom, `scrollback_len` = top).
